requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
Django==3.2.7
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-paginate==0.8.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
image==1.5.33
ipaddress==1.0.23
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
monero==0.8.2
numpy==1.21.2
Pillow==8.3.2
pycparser==2.20
pysha3==1.0.2
pytz==2021.1
qrcode==7.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
urllib3==1.26.6
varint==1.0.2
Werkzeug==2.0.1
WTForms==2.3.3
